// api-gateway/src/interfaces/auth_service.ts
export interface User {
  id: string;
  username: string;
  picture: string;
  email: string;
  password: string;
  role: string;
  isVerified: boolean;
  isTwoFactorEnabled: boolean;
  method: string;
}